splits IsDirWritable
