+++ internal/environment/default_test.go
@@ -1,6 +1,9 @@
 // Copyright 2022 Juan Pablo Tosso and the OWASP Coraza contributors
 // SPDX-License-Identifier: Apache-2.0
 
+//go:build !no_fs_access
+// +build !no_fs_access
+
 package environment
 
 import (
@@ -18,28 +21,20 @@ func TestFSCheck(t *testing.T) {
 		expectError   bool
 	}{
 		{
-			name:          "No access to FS",
-			hasAccessToFS: false,
-			expectError:   false,
+			name:        "Has access to FS, non-existent dir",
+			tmpDir:      "/non-existent-dir",
+			expectError: true,
 		},
 		{
-			name:          "Has access to FS, non-existent dir",
-			hasAccessToFS: true,
-			tmpDir:        "/non-existent-dir",
-			expectError:   true,
-		},
-		{
-			name:          "Has access to FS, existent dir",
-			hasAccessToFS: true,
-			tmpDir:        os.TempDir(),
-			expectError:   false,
+			name:        "Has access to FS, existent dir",
+			tmpDir:      os.TempDir(),
+			expectError: false,
 		},
 	}
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			os.Setenv("TMPDIR", tc.tmpDir)
-			HasAccessToFS = tc.hasAccessToFS
 			err := IsDirWritable(debuglog.Default(), tc.tmpDir)
 			if tc.expectError && err == nil {
 				t.Fatal("expected error, got nil")
