+++ waf.go
@@ -34,8 +34,10 @@ func NewWAF(config WAFConfig) (WAF, error) {
 
 	waf := corazawaf.NewWAF()
 
-	if err := environment.IsDirWritable(waf.Logger, waf.TmpDir); err != nil {
-		return nil, fmt.Errorf("filesystem access check: %w. Use 'no_fs_access' build tag, if not available", err)
+	if environment.HasAccessToFS {
+		if err := environment.IsDirWritable(waf.Logger, waf.TmpDir); err != nil {
+			return nil, fmt.Errorf("filesystem access check: %w. Use 'no_fs_access' build tag, if not available", err)
+		}
 	}
 
 	if c.debugLogger != nil {
